pyproject.toml
[project]
name = "hanz"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "elevenlabs>=1.0.0",
    "google-genai>=1.0.0",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.0",
]
